Extend SR benchmarks to 80 fields
Using the unsafe API allows us to do so.

This also fixes the measurement tools to close each file after we
process it, to avoid running out of file handles.

We only show -ddump-simpl for superrecord to avoid the core size
measurements taking to look. The full benchmark suite now takes roughly
25 minutes on my machine.

+++ large-records-benchmarks/report/update.sh
@@ -5,7 +5,11 @@ set -e
 #
 # Run this from the report/ directory.
 #
-# To benchmark only coresize, timing, or time, set the BENCH env var.
+# To benchmark only "coresize", "timing", or "runtime", set the BENCH env var.
+# To benchmark only only a specific target, set TARGET.
+#
+# NOTE: If benchmarking only a single target, you will need to manually remove
+# old lines from runtime.csv.
 #
 
 
@@ -14,12 +18,24 @@ set -e
 if [[ "$BENCH" == "" || "$BENCH" == "coresize" ]]
 then
 
-  cabal build bench-before       --flags=+profile-coresize
-  cabal build bench-after        --flags=+profile-coresize
-  cabal build bench-experiments  --flags=+profile-coresize
-  cabal build bench-typelet      --flags=+profile-coresize
-  cabal build bench-large-anon   --flags=+profile-coresize
-  cabal build bench-superrecord  --flags=+profile-coresize
+  if [[ "$TARGET" == "" || "$TARGET" == "before" ]]; then
+    cabal build bench-before --flags=+profile-coresize
+  fi
+  if [[ "$TARGET" == "" || "$TARGET" == "after" ]]; then
+    cabal build bench-after --flags=+profile-coresize
+  fi
+  if [[ "$TARGET" == "" || "$TARGET" == "experiments" ]]; then
+    cabal build bench-experiments --flags=+profile-coresize
+  fi
+  if [[ "$TARGET" == "" || "$TARGET" == "typelet" ]]; then
+    cabal build bench-typelet --flags=+profile-coresize
+  fi
+  if [[ "$TARGET" == "" || "$TARGET" == "large-anon" ]]; then
+    cabal build bench-large-anon --flags=+profile-coresize
+  fi
+  if [[ "$TARGET" == "" || "$TARGET" == "superrecord" ]]; then
+    cabal build bench-superrecord --flags=+profile-coresize
+  fi
 
   cabal run parse-coresize -- \
     --dist ../../dist-newstyle \
@@ -33,12 +49,24 @@ fi
 if [[ "$BENCH" == "" || "$BENCH" == "timing" ]]
 then
 
-  cabal build bench-before       --flags=+profile-timing
-  cabal build bench-after        --flags=+profile-timing
-  cabal build bench-experiments  --flags=+profile-timing
-  cabal build bench-typelet      --flags=+profile-timing
-  cabal build bench-large-anon   --flags=+profile-timing
-  cabal build bench-superrecord  --flags=+profile-timing
+  if [[ "$TARGET" == "" || "$TARGET" == "before" ]]; then
+    cabal build bench-before --flags=+profile-timing
+  fi
+  if [[ "$TARGET" == "" || "$TARGET" == "after" ]]; then
+    cabal build bench-after --flags=+profile-timing
+  fi
+  if [[ "$TARGET" == "" || "$TARGET" == "experiments" ]]; then
+    cabal build bench-experiments --flags=+profile-timing
+  fi
+  if [[ "$TARGET" == "" || "$TARGET" == "typelet" ]]; then
+    cabal build bench-typelet --flags=+profile-timing
+  fi
+  if [[ "$TARGET" == "" || "$TARGET" == "large-anon" ]]; then
+    cabal build bench-large-anon --flags=+profile-timing
+  fi
+  if [[ "$TARGET" == "" || "$TARGET" == "superrecord" ]]; then
+    cabal build bench-superrecord --flags=+profile-timing
+  fi
 
   cabal run parse-timing -- \
     --dist ../../dist-newstyle \
@@ -50,13 +78,19 @@ fi
 
 ## Runtime
 
-rm -f runtime.csv
-
 if [[ "$BENCH" == "" || "$BENCH" == "runtime" ]]
 then
 
-  cabal run --flags=+profile-runtime bench-large-anon  -- --csv runtime.csv
-  cabal run --flags=+profile-runtime bench-superrecord -- --csv runtime.csv
+  if [[ "$TARGET" == "" ]]; then
+    rm -f runtime.csv
+  fi
+
+  if [[ "$TARGET" == "" || "$TARGET" == "large-anon" ]]; then
+    cabal run --flags=+profile-runtime bench-large-anon  -- --csv runtime.csv
+  fi
+  if [[ "$TARGET" == "" || "$TARGET" == "superrecord" ]]; then
+    cabal run --flags=+profile-runtime bench-superrecord -- --csv runtime.csv
+  fi
 
 fi
